Refactor: extract whitelist query into single file
