+++ src/data/graphql/queries.graphql
@@ -102,15 +102,6 @@ query UserAddress($name: String!) {
   userAddress(name: $name) @client
 }
 
-query WhitelistedUsers($colonyAddress: String!) {
-   whitelistedUsers(colonyAddress: $colonyAddress) @client {
-    id
-    profile {
-      walletAddress
-    }
-   }
-}
-
 query TokenBalancesForDomains($colonyAddress: String!, $tokenAddresses: [String!]!, $domainIds: [Int!]) {
   tokens(addresses: $tokenAddresses) @client {
     id
@@ -667,32 +658,5 @@ query ColonyHistoricRoles($colonyAddress: String!, $blockNumber: Int!) {
       domainId
       roles
     }
-  }
-}
-
-query WhitelistAgreement($agreementHash: String!) {
-  whitelistAgreement(agreementHash: $agreementHash) @client
-}
-
-query WhitelistAgreementHash($colonyAddress: String!) {
-  whitelistAgreementHash(colonyAddress: $colonyAddress) @client
-}
-
-query HasKycPolicy($colonyAddress: String!) {
-  hasKycPolicy(colonyAddress: $colonyAddress) @client
-}
-
-query WhitelistPolicy($colonyAddress: String!) {
-  whitelistPolicy(colonyAddress: $colonyAddress) @client {
-    kycRequired,
-    agreementRequired,
-  }
-}
-
-query UserWhitelistStatus($colonyAddress: String!, $userAddress: String!) {
-  userWhitelistStatus(colonyAddress: $colonyAddress, userAddress: $userAddress) @client {
-    userIsApproved,
-    userIsWhitelisted,
-    userSignedAgreement,
   }
 }
