+++ src/data/graphql/queries/extensions.graphql
@@ -102,14 +102,3 @@ query SubgraphExtensionEvents($colonyAddress: String!, $extensionAddress: String
     }
   }
 }
-
-query SubgraphKYCAddresses($extensionAddress: String!) {
-    kycaddresses(
-      where: {
-        extension_contains: $extensionAddress,
-        status: true
-      }
-    ) {
-    walletAddress: id
-  }
-}
